Weather app bar separate to widget.
