+++ lib/common/widgets/components/item_views/next_days_weather_item_views.dart
@@ -3,7 +3,6 @@ import 'package:feature_first/common/widgets/components/background_card.dart';
 import 'package:feature_first/common/widgets/icon_image_view.dart';
 import 'package:feature_first/utils/utils.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class NextDaysWeatherItemView extends StatelessWidget {
